+++ keratin-log/src/bin/keratin_bench_auto_batch.rs
@@ -21,6 +21,7 @@ fn main() {
         min_fsync_interval_ms: 0,
         flush_target_bytes: 48 * 1024 * 1024,
         default_durability: KDurability::AfterFsync,
+        writer_buffer_factor: 128,
         force_recovery_scan: false,
         tail_cache_bytes: 0,
         segment_preallocate_bytes: 0,